requirements.txt
ctransformers==0.2.5
sentence-transformers==3.2.1
pinecone-client
langchain==0.0.225
flask
pypdf
python-dotenv




-e .